Rewrote/refactored base/ELPV configuration, datasets, and schemas.
Made separate subpackage for base classes regarding configuration, schemas, and datasets from their concrete implementation.
