+++ tests/test_datasets.py
@@ -4,9 +4,8 @@ import numpy as np
 import pytest as pytest
 from PIL import Image
 
-from tmap_defectdetector.dataset_downloaders import DatasetDownloaderELPV
-from tmap_defectdetector.datasets import AbstractDataSet, ImageDataSetELPV
-from tmap_defectdetector.samples import SampleLabelsELPV
+from tmap_defectdetector.dataset.datasets import ImageDataSetELPV
+from tmap_defectdetector.dataset.downloaders import DatasetDownloaderELPV
 
 
 def download_elpv_dataset_and_return_downloader():
@@ -19,15 +18,15 @@ def image_dataset_ELPV() -> ImageDataSetELPV:
     raise NotImplementedError()
 
 
-@pytest.fixture
-def abstract_dset():
-    return AbstractDataSet()
+# @pytest.fixture
+# def abstract_dset():
+#     return AbstractDataSet()
 
 
-@pytest.fixture
-def sample_labels_ELPV() -> SampleLabelsELPV:
-    downloader = download_elpv_dataset_and_return_downloader()
-    return SampleLabelsELPV.from_csv(csv_path=downloader.label_paths[0])
+# @pytest.fixture
+# def sample_labels_ELPV() -> SampleLabelsELPV:
+#     downloader = download_elpv_dataset_and_return_downloader()
+#     return SampleLabelsELPV.from_csv(csv_path=downloader.label_paths[0])
 
 
 def _gen_img(img_width: int = 255, img_height: int = 255) -> Image:
@@ -35,13 +34,13 @@ def _gen_img(img_width: int = 255, img_height: int = 255) -> Image:
     return Image.fromarray(np.uint8(np.random.rand(img_width, img_height))).convert("RGB")
 
 
-class TestAbstractDataSet:
-    def test_amplify_data(self):
+# class TestAbstractDataSet:
+#     def test_amplify_data(self):
+#
+#         with pytest.raises(NotImplementedError):
+#             abstract_dset.amplify_data()
 
-        with pytest.raises(NotImplementedError):
-            abstract_dset.amplify_data()
-
-
-class TestImageDataset(AbstractDataSet):
-
-    pass
+#
+# class TestImageDataset(AbstractDataSet):
+#
+#     pass
