+++ tmap_defectdetector/dataset/base/__init__.py
@@ -1,0 +1,6 @@
+"""
+This subpackage contains the baseclasses required to ipmlement downloading and loading (defect)
+specific (image) datasets, such as the schemas (column name -> column type mappings),
+expected data amplification methods, etc.
+"""
+from __future__ import annotations
