+++ tmap_defectdetector/dataset/base/downloaders_base.py
@@ -1,7 +1,4 @@
-"""
-Contains abstract baseclasses for downloading datasets,
-as well as implementations thereof for downloading specific datasets.
-"""
+"""Defines baseclasses for downloading datasets."""
 from __future__ import annotations
 
 import os
@@ -12,10 +9,8 @@ from typing import Iterable, Callable
 from git import Repo, Remote
 from tqdm import tqdm
 
-from tmap_defectdetector.config.paths import DIR_DATASETS
 from tmap_defectdetector.logger import log
-
-from tmap_defectdetector.datasets import AbstractDataSet
+from tmap_defectdetector.pathconfig.paths import DIR_DATASETS
 
 
 class AbstractDatasetDownloader(ABC):
@@ -28,11 +23,6 @@ class AbstractDatasetDownloader(ABC):
     @abstractmethod
     def download(self) -> None:
         """This method should handle the download portion for the dataset."""
-        pass
-
-    @abstractmethod
-    def load(self) -> AbstractDataSet:
-        """This method should handle the loading of the dataset into a common format."""
         pass
 
 
@@ -90,11 +80,6 @@ class DatasetDownloaderGit(AbstractDatasetDownloader):
         """URL pointing to repository, cannot be changed; instantiate a new object instead."""
         return self._repo_url
 
-    def load(self) -> AbstractDataSet:
-        raise NotImplementedError(
-            "Conversion of this dataset to 'datasets.DataSet' format not yet implemented."
-        )
-
     def download(
         self,
         tgt_rootdir: Path = DIR_DATASETS,
@@ -141,7 +126,7 @@ class DatasetDownloaderGit(AbstractDatasetDownloader):
         return repo, origin
 
     def _fetch(self, origin: Remote) -> Remote:
-        """Used internally. Fetches repository from a Remote"""
+        """Used internally. Fetches repository from a Remote."""
         log.info(f"Fetching dataset from remote git (url={self.repo_url!r})...")
         origin.fetch()
         log.info("Dataset fetch from remote git complete!")
@@ -202,62 +187,8 @@ class DatasetDownloaderGit(AbstractDatasetDownloader):
             for file in tqdm(
                 sample_dir.rglob("*.*"),
                 desc=f"Finding {self.DEFAULT_DATASET_UNITS}...",
-                unit=self.DEFAULT_DATASET_UNITS,
+                unit=f" {self.DEFAULT_DATASET_UNITS}",
             ):
                 if filechecker_function(file):
                     data_paths.append(file)
         return data_paths
-
-
-class DatasetDownloaderELPV(DatasetDownloaderGit):
-    # Note that there is probably a far better more generalized approach than defining a class to
-    # download/preprocess every dataset.
-    # However, given that we probably don't have to use that many datasets and it would likely take to much
-    # time to implement a generalizable way to do this right now, the choice was made to do it like this.
-
-    # Default repository name for this DatasetDownloader subclass
-    DEFAULT_REPO_NAME: str = "dataset-elpv"
-
-    # Description of the type of sample data used for e.g. progress bars,
-    # should not impact functionality
-    DEFAULT_DATASET_UNITS: str = "image samples"
-
-    def __init__(
-        self,
-        dataset_name: str = DEFAULT_REPO_NAME,
-        relative_sample_dir_paths: Iterable[os.PathLike] = (Path("images"),),
-        relative_label_file_paths: Iterable[os.PathLike] = (Path("labels.csv"),),
-        repo_url: str = "https://github.com/jrbergen/elpv-dataset",
-    ):
-        """
-        Downloads datasets from https://github.com/jrbergen/elpv-dataset
-        which contains example datasets forked from https://github.com/zae-bayern/elpv-dataset
-
-        :param repo_url: (optional) url pointing to the remote repository for the ELPV dataset
-            (Defaults to "https://github.com/jrbergen/elpv-dataset").
-        :param dataset_name: (optional str) name for this dataset
-            (defaults to DatasetDownloaderELPV.DEFAULT_REPO_NAME).
-        :param relative_sample_dir_paths: (optional) iterable containing one or more paths
-            to director[y|ies] containing the training data relative to the root repository directory.
-            e.g. if the repository is called "dataset_repo" and the directories containing training images
-            are in "dataset_repo/training_data/image_set1" and "dataset_repo/training_data/image_set2",
-            then you could pass ["dataset_repo/training_data/image_set1", "dataset_repo/training_data/image_set2"]
-            to this argument. Defaults to "images" (==dir w/ training data the ELPV dataset).
-        :param relative_label_file_paths: (optional) iterable containing one or more paths
-            to a file containing the labels (e.g. a CSV file) relative to the root repository directory.
-            e.g. if the repository is called "dataset_repo" and the 'labels' file is in
-            "dataset_repo/labels/labels_set1.csv", then you should pass ["labels/image_set1.csv"]
-            to this argument. Defaults to "labels.csv" (==file w/ labels for the ELPV dataset).
-            Warning: will assume every file in this relative directory is a file containing labels.
-
-        """
-        super().__init__(
-            repo_url=repo_url,
-            dataset_name=dataset_name,
-            relative_sample_dir_paths=relative_sample_dir_paths,
-            relative_label_file_paths=relative_label_file_paths,
-        )
-
-        # The image loading logic which was here before is now part of the ImageDataSetELPV class.
-        # We may want to alter this class later (e.g. remove it entirely and find a better way to
-        # make its DataSetDownloader parent class suited to donwloading the ELPV dataset specifically).
