+++ tmap_defectdetector/dataset/base/schemas_base.py
@@ -1,37 +1,18 @@
 """
-Contains classes for handling data schemas (column name- and type specifications),
-and utility functions to e.g. combine them or build an empty DataFrame out of a schema.
-This file also contains baseclasses for dataset configuration.
+Contains baseclasses with which to build schemas (i.e. column name and type mappings) for new datasets.
 """
 
 from __future__ import annotations
 
-import inspect
-import os
 import string
 import warnings
-from abc import abstractmethod, ABC
 from dataclasses import dataclass
-from functools import cached_property
-
-
-from pathlib import Path
-from typing import (
-    Iterable,
-    TypeAlias,
-    ClassVar,
-    TypeVar,
-    Iterator,
-    cast,
-    Mapping,
-    Collection,
-)
+from typing import TypeAlias, Mapping, Iterator, TypeVar, Type
 
 import numpy as np
-from pandas import DataFrame
 import pandas as pd
+from pandas import DataFrame
 from pandas.core.dtypes.base import ExtensionDtype
-
 
 ColType: TypeAlias = ExtensionDtype | np.dtype | type | str
 
@@ -63,12 +44,14 @@ class SchemaEntry:
     def __init__(self, name: ColName | str, type_: ColType, docstring: str = ""):
         """
         Dataclass storing a single column name + data type combination (i.e. a single schema entry).
+        Supports addition of documentation strings / descriptions to the entries.
 
         :param name: ColName or string instance representing column name.
         :param type_: column type for this schema entry (ExtensionDtype | np.dtype | type | str)
         :param docstring: (optional) doctsring describing this schema entry and/or its purpose.
         """
-        self._name, self._type = "", ""
+        self._name: ColName = ColName("")
+        self._type: ColType = ""
         self.name: ColName = name if type(name) is ColName else ColName(str(name))
         self.type: ColType = type_
         self.docstring: str = docstring
@@ -124,6 +107,10 @@ class SchemaEntry:
         return f"{type(self).__name__}(name={self._name!r}, type={typedesc}, docstring={repr(self.docstring)})"
 
 
+SchemaDerived = TypeVar("SchemaDerived")
+"""Type variable to indicate some variable schema derived type."""
+
+
 @dataclass
 class ColSchema:
     def __init__(self, colentries: Mapping[str, SchemaEntry]):
@@ -134,14 +121,14 @@ class ColSchema:
         :param colentries: mapping/dictionary of attribute names as keys
             with associated SchemaEntry objects as values.
         """
-        self.add_entries(**colentries)
+        self.add_entries(colentries)
 
-    def add_entries(self, **colentries) -> None:
+    def add_entries(self, colentries: Mapping[str, SchemaEntry]) -> None:
         """
         Adds SchemaEntry entries to this schema
         (duplicate column names or existing attribute names are not allowed).
 
-        :param **colentries: keyword arguments with SchemaEntry objects as values
+        :param colentries: a mapping of attribute names to SchemaEntry objects.
         """
         encountered_attrnames = set()
         encountered_colnames = set()
@@ -249,13 +236,23 @@ class ColSchema:
 
     def __add__(self, other: ColSchema) -> ColSchema:
         """
-        Adds two ColSchema instances (addition is equivalent to union for this class).
+        Adds/merges two ColSchema instances (addition is equivalent to union for this class).
+
+        :raises ValueError: if non-identical SchemaEntry members of the schemas have identical column names.
+        :raises ValueError: if non-identical SchemaEntry members of the schemas have identical attribute names.
+        """
+        return ColSchema(colentries=self._merged_schemas_to_attrdict(other))
+
+    def _merged_schemas_to_attrdict(self, other: ColSchema):
+        """
+        Adds/merges two ColSchema instances (addition is equivalent to union for this class).
 
         :raises ValueError: if non-identical SchemaEntry members of the schemas have identical column names.
         :raises ValueError: if non-identical SchemaEntry members of the schemas have identical attribute names.
         """
         if not isinstance(other, ColSchema):
             return NotImplemented
+
         schema_dict_self = {k: v for k, v in vars(self).items() if isinstance(v, SchemaEntry)}
         schema_dict_other = {k: v for k, v in vars(other).items() if isinstance(v, SchemaEntry)}
         for attr_self, entry_self in schema_dict_self.items():
@@ -268,10 +265,24 @@ class ColSchema:
                         f"different values, to prevent accidental overwrites."
                     )
         schema_dict_self.update(**schema_dict_other)
-        return ColSchema(**schema_dict_self)  # type: ignore
+        return schema_dict_self
+
+    def combine_with_schema(
+        self, other_schema: ColSchema, target_schema_type: Type[SchemaDerived]
+    ) -> SchemaDerived:
+        new_schema_dict: dict[str, SchemaEntry] = self._merged_schemas_to_attrdict(other_schema)
+        if not issubclass(target_schema_type, ColSchema):
+            raise TypeError(f"Target schema must be a subclass of {target_schema_type.__name__}.")
+        # !NOTE: This 'SchemaDerived' typevar approach seems wrong; find how this should be done.
+        target_schema_type: Type[ColSchema]  # type: ignore
+        return target_schema_type(colentries=new_schema_dict)  # type: ignore
 
     def __repr__(self) -> str:
-        return f"{ColSchema.__name__}({', '.join(f'{k}={v}' for k, v in vars(self).items() if isinstance(v, SchemaEntry))})"
+        return (
+            f"{ColSchema.__name__}("
+            + ", ".join(f"{k}={v}" for k, v in vars(self).items() if isinstance(v, SchemaEntry))
+            + ")"
+        )
 
 
 @dataclass(repr=False)
@@ -296,171 +307,44 @@ class ColSchemaDefectData(ColSchema):
 ColSchemaDefectDataType = TypeVar("ColSchemaDefectDataType", bound=ColSchemaDefectData)
 
 
-class DataSetConfig(ABC):
+@dataclass(repr=False)
+class SchemaSamplesImageData(ColSchemaDefectData):
+    """
+    Used to specify the schema (column names and data types) for a
+    defect dataset/labelset/sampleset pertaining image samples specifically.
+    """
 
-    _CACHED_PROPERTIES: tuple[str, ...] = ("full_dataset", "label_data", "sample_data")
-    """Properties which need to be (re)loaded if an attribute changes change."""
-
-    def __init__(
-        self,
-        sample_dirs: os.PathLike | Collection[os.PathLike],
-        sample_col_schema: ColSchemaDefectDataType,
-        label_path: os.PathLike,
-        label_col_schema: ColSchemaDefectDataType,
-        sample_type_desc: str = "sample",
-    ):
-        """
-        Provides ways to load a training dataset's samples and labels into a DataFrame.
-
-        :param sample_dirs: :param sample_dirs: One ore more path-like object(s)
-            pointing to a directory with sample files.
-        :param sample_col_schema: ColSchemaDefectData object representing
-            column schema (column names + dtypes) for the DataFrame to
-            be created which will contain the samples.
-        :param label_path: One ore more path-like object(s)
-            pointing to corresponding label files.
-        :param label_col_schema: ColSchemaDefectData object representing
-            column schema (column names + dtypes) for the DataFrame to
-            be created which will contain the samples.
-        :param sample_type_desc: (optional) description for this kind
-            of sample (default = "sample").
-        """
-        self.sample_dirs: list[Path] = (
-            list(Path(d) for d in sample_dirs)
-            if isinstance(sample_dirs, Iterable)
-            else [Path(sample_dirs)]
-        )
-        self.label_path = Path(label_path)
-
-        self.sample_col_schema: ColSchema = sample_col_schema
-        self.label_col_schema: ColSchema = label_col_schema
-
-        self.sample_type_desc: str = sample_type_desc
-
-    @classmethod
-    def file_is_sample(cls, file: Path) -> bool:
-        """
-        Function used to check if a file can be identified as / considered a sample for this dataset.
-        Defaults to just checking whether a Path indeed points to a file. Override in extending classes
-        to perform more comprehensive checks.
-
-        :param file: Path to (potential) sample file.
-        """
-        return file.is_file()
-
-    @cached_property
-    @abstractmethod
-    def full_dataset(self) -> DataFrame:
-        """
-        Merges sample and label DataFrames into one coherent whole.
-        _Must_ be implemented by subclasses.
-        """
-        ...
-
-    @cached_property
-    @abstractmethod
-    def label_data(self) -> pd.DataFrame:
-        """
-        Provides way to load labels for a specific dataset into
-        a DataFrame using the label path passed to the constructor.
-        _Must_ be implemented by subclasses.
-        """
-        ...
-
-    @cached_property
-    @abstractmethod
-    def sample_data(self) -> DataFrame:
-        """
-        Provides way to load samples for a specific dataset into
-        a DataFrame using the label path passed to the constructor.
-        _Must_ be implemented by subclasses.
-        """
-        ...
-
-    @property
-    def column_schema(self) -> ColSchema:
-        """
-        ColSchema object containing all column names and data types
-        for the dataframe containing samples + labels.
-        """
-        return self.label_col_schema | self.sample_col_schema
-
-    @property
-    def label_colnames(self) -> tuple[str, ...]:
-        """Retrieve column names for the label data."""
-        return tuple(self.label_col_schema.keys())  # type: ignore
-
-    @property
-    def sample_colnames(self) -> tuple[str, ...]:
-        """Retrieve column names for the sample data."""
-        return tuple(self.sample_col_schema.keys())  # type: ignore
-
-    @property
-    def label_coltypes(self) -> tuple[str, ...]:
-        """Retrieve column datatypes for the label data."""
-        return tuple(self.label_col_schema.values())  # type: ignore
-
-    @property
-    def sample_coltypes(self) -> tuple[str, ...]:
-        """Retrieve column datatypes for the sample data."""
-        return tuple(self.sample_col_schema.values())  # type: ignore
-
-    def __setattr__(self, key, value):
-        # Override setattr such that cached properties are reset if the relevant attributes have changed
-        self.__dict__[key] = value
-        for cached_attrname in type(self)._CACHED_PROPERTIES:
-            if cached_attrname in self.__dict__:
-                self.__dict__.pop(cached_attrname, None)
-
-    def __repr__(self) -> str:
-        return f"""{type(self).__name__}({", ".join(f'{k}={v}' for k, v in vars(self).items())})"""
+    MIRROR_AXIS: SchemaEntry = SchemaEntry(
+        "MIRROR_AXIS",
+        np.uint8,
+        docstring="Number representing mirrored state w.r.t original image.",
+    )
+    ROT_DEG: SchemaEntry = SchemaEntry(
+        "ROT_DEG", np.int16, docstring="Rotation of image in degrees w.r.t. original image."
+    )
+    TRANSL_X: SchemaEntry = SchemaEntry(
+        "TRANSL_X", np.int16, docstring="Translation of Y-pixel positions w.r.t. original image."
+    )
+    TRANSL_Y: SchemaEntry = SchemaEntry(
+        "TRANSL_Y", np.int16, docstring="Translation of X-pixel positions w.r.t. original image."
+    )
+    SAMPLE_PATH: SchemaEntry = SchemaEntry(
+        "SAMPLE_PATH",
+        str,
+        docstring="Full path of the file which stores the data labels, as string.",
+    )
+    SAMPLE: SchemaEntry = SchemaEntry(
+        "SAMPLE",
+        object,  # should be an np.ndarray but use 'object' as type for compatibility with pandas DataFrame.
+        docstring="A numpy array representing the sample image.",
+    )
 
 
-class DefectDetectionDataSet:
+@dataclass(repr=False)
+class SchemaLabels(ColSchemaDefectData):
+    """Specifies schema for label data."""
 
-    DEFAULT_DATASET_UNITS: ClassVar[str] = " samples"
 
-    def __init__(self, sample_and_label_data: DataFrame):
-        """Baseclass for a generic dataset used for training a defect detection model."""
-        self._data_original: DataFrame = sample_and_label_data
-        self._data_filtered: DataFrame = sample_and_label_data.copy()
-
-    @classmethod
-    def from_dataset_config(cls, dataset_cfg: DataSetConfig) -> DefectDetectionDataSet:
-        """
-        Instantiates a DefectDetectionDataSet (derived) object based on a dataset configuration
-        (i.e. DataSetConfig object).
-
-        :param dataset_cfg: DataSetConfig derived object specifying
-            how/where to read in the data labels and samples.
-        """
-        return cls(sample_and_label_data=dataset_cfg.full_dataset)
-
-    def amplify_data(self):
-        """
-        Performs operations which effectively increase the dataset size
-        as to reduce overfitting problems / allow for a more generalizable
-        model.
-        This can be done by e.g. by mirroring, rotating, translating,
-        or applying filters in case the training data comprises images.
-        Subclasses should implement this method.
-        """
-        raise NotImplementedError(
-            f"method {inspect.currentframe().f_code.co_name} not implemented for baseclass."
-        )
-
-    def reset(self) -> None:
-        """Resets dataset to initialization state (before any filters were applied)."""
-        self._data_filtered = self._data_original
-
-    @property
-    def data(self) -> DataFrame:
-        """
-        Returns a dataframe w/ _filtered_ samples and labels if any filter has been applied.
-        To reset to the unfiltered version, call the `reset()` method.
-        Note that this is a read-only property.
-        """
-        return self._data_filtered
-
-    def __repr__(self) -> str:
-        return f"{type(self).__name__}(data={self.data})"
+@dataclass(repr=False)
+class SchemaFullImageData(SchemaLabels, SchemaSamplesImageData):
+    """Specifies schema for an image dataset's label _and_ sample data."""
