+++ tmap_defectdetector/main.py
@@ -1,13 +1,13 @@
 """The main file containing the program's entrypoint."""
 from __future__ import annotations
 
-from pathlib import Path
-
 from tmap_defectdetector.compatibility_checks import version_check
-from tmap_defectdetector.dataset_downloaders import DatasetDownloaderELPV
+from tmap_defectdetector.dataset.datasets import ImageDataSetELPV
+from tmap_defectdetector.dataset.downloaders import DatasetDownloaderELPV
 
-from tmap_defectdetector.datasets import ImageDataSetELPV
-from tmap_defectdetector.image_helpers import file_is_image
+from tmap_defectdetector.dataset.dataset_configs import DataSetConfigELPV
+from tmap_defectdetector.dataset.schemas import SchemaLabelsELPV
+
 from tmap_defectdetector.logger import log
 
 
@@ -15,27 +15,26 @@ def cli():
     ...
 
 
-def main():
-    version_check()
-
+def example_elpv():
     # Initialize the dataset downloader and download the ELPV dataset from its git repository.
     downloader = DatasetDownloaderELPV()
     downloader.download()  # The dataset is downloaded to %LOCALAPPDATA%/.tmapdd/datasets/dataset-elpv/ (on Windows)
 
-    # We first find the image paths in the image dataset directories and accept the files
-    # which comply with our anonymous (lambda) function check.
-    sample_files = downloader.get_data_files(
-        filechecker_function=lambda p: Path.is_file(p) and file_is_image(p)
-    )
-    label_files = downloader.label_paths
+    # Initialize/load the ELPV dataset using the ELPV dataset configuration.
+    elpv_dataset_config = DataSetConfigELPV()
+    dataset = ImageDataSetELPV(dataset_config=elpv_dataset_config)
 
-    # Now we know where the label path(s) and the data (image file) paths are, we can construct the ImageDataSet.
-    dataset = ImageDataSetELPV.from_paths(data_paths=sample_files, label_paths=label_files)
+    # Filter dataset -> use only the polycrystalline solarpanels w/ type 'poly'.
+    dataset.filter(query=f"{SchemaLabelsELPV().TYPE.name}=='poly'")
 
     # Here comes the preprocessing step (we could e.g. make a ImageDataSetPreProcessor class/function or perhaps
     # put preprocessing methods in the ImageDataSet class itself later.
     dataset.amplify_data()
 
+
+def main():
+    version_check()
+    example_elpv()
     log.info("All done!")
