+++ tmap_defectdetector/pathconfig/__init__.py
